moved from plugins/ to domain/model
