+++ src/main/kotlin/io/sebi/domain/model/TreeNode.kt
@@ -1,4 +1,4 @@
-package io.sebi.plugins
+package io.sebi.domain.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
